Fix LockFileUtilities.IsLockFileStillValid tests
